feat : 가게 상세 페이지 찜, 리뷰 점수, 대표이미지 추가

+++ src/main/java/yu/cse/odiga/store/dto/ReviewRegisterDto.java
@@ -8,6 +8,6 @@ import org.springframework.web.multipart.MultipartFile;
 @Getter
 public class ReviewRegisterDto {
     private String content;
-    private double rating;
+    private int rating;
     private MultipartFile image;
 }
